move kube_api configs and cleanup
