+++ airflow_kubernetes_job_operator/kube_api/collections.py
@@ -1,6 +1,6 @@
 import re
 from enum import Enum
-from typing import Callable, Dict, List
+from typing import Callable, List
 from airflow_kubernetes_job_operator.kube_api.utils import not_empty_string
